Update contact email
The contact email was updated from ‘integrationsupport@acc.co.nz’ to ‘digitaloperations@acc.co.nz’ in the open API configuration. The swagger's URL for the Integrated Care Pathways endpoint was also modified to point to a new location. The contact email was similarly updated in the 'index.md' file for feedback communications.