Add STarT-C questionnaire for Traumatic Brain Injury patients
This commit introduces the "STarT-C" questionnaire which is used for Traumatic Brain Injury patients. It also creates defined response options for individual questions in the questionnaire, such as 'Agree' and 'Disagree' as well as the ordinal values for the overall STarT-C concussion symptoms. A change is made to the 'capstmt.fsh' file and a few additional aliases are added.