Refactor patient referral IDT instance and add intent and category
Refactored the patient referral IDT instance to fix a typo in the instance name and added the intent and category fields. The intent field is set to #order and the category field is set to SCT code #770656003. This change improves the accuracy and completeness of the patient referral IDT instance.